+++ src/gcdatastore.ts
@@ -1,7 +1,7 @@
 import {Session} from '@shopify/shopify-api';
 import {SessionStorage} from '@shopify/shopify-app-session-storage';
 
-export class MemorySessionStorage implements SessionStorage {
+export class GCDatastoreSessioNStorage implements SessionStorage {
   private sessions: {[id: string]: Session} = {};
 
   public async storeSession(session: Session): Promise<boolean> {